--- author.json
@@ -1,1 +1,0 @@
-[{"id":730,"first_name":"dorian","last_name":"urem","publish_date":"2000-02-02"},{"id":12,"first_name":"d","last_name":"u"}]
