--- books.json
@@ -1,1 +1,0 @@
-[{"id":175,"publisher":"kaka","archived":true,"cover_state":"bad","publish_date":"2000-02-02","label_id":365},{"id":961,"publisher":"d","archived":true,"cover_state":"bad","publish_date":"0200-02-02","label_id":234}]
