+++ game.rb
@@ -1,5 +1,6 @@
 require './item'
 require 'date'
+
 class Game < Item
   attr_accessor :multiplayer, :last_played_at_date
 
@@ -7,7 +8,8 @@ class Game < Item
     super(publish_date)
 
     @multiplayer = multiplayer
-    @last_played_at_date = last_played_at_date
+    @last_played_at_date = Date.parse(last_played_at_date)
+    move_to_archive
   end
 
   def can_be_archived?
