--- games.json
@@ -1,1 +1,0 @@
-[{"id":789,"multiplayer":true,"archived":false,"author_id":730,"last_played_at_date":"2000-2-22","publish_date":"2000-02-02"},{"id":846,"multiplayer":true,"archived":false,"author_id":12,"last_played_at_date":"2000-2-2","publish_date":"2000-02-02"}]
