--- genres.json
@@ -1,1 +1,0 @@
-[{"id":157,"name":"comedy"},{"id":673,"name":"no"}]
