+++ main.rb
@@ -1,28 +1,30 @@
 require './app'
+require './list'
 
 def main
-  app = App.new
+  app = AppAdd.new
+  list = AppList.new
   puts 'Welcome to Catalog of my things App!'
   option = display_list
-  list_items(app, option)
+  list_items(list, option)
   add_items(app, option)
   main unless option.zero?
 end
 
-def list_items(app, option)
+def list_items(list, option)
   case option
   when 1
-    app.list_all_books
+    list.list_all_books
   when 2
-    app.list_all_musics
+    list.list_all_musics
   when 3
-    app.list_all_games
+    list.list_all_games
   when 4
-    app.list_all_genres
+    list.list_all_genres
   when 5
-    app.list_all_labels
+    list.list_all_labels
   when 6
-    app.list_all_authors
+    list.list_all_authors
   end
 end
