--- musics.json
@@ -1,1 +1,0 @@
-[{"id":886,"on_spotify":"true","archived":false,"genre_id":157,"publish_date":"2000-02-22"},{"id":713,"on_spotify":"true","archived":false,"genre_id":673,"publish_date":"2000-02-02"}]
