+++ spec/author_spec.rb
@@ -3,7 +3,7 @@ require_relative('spec_helper')
 describe Author do
   before :all do
     @game = Game.new(true, '2000-10-10', '1998-10-11')
-    @author = Author.new('Alfred', 'Kinyua', '2022-10-10')
+    @author = Author.new('Alfred', 'Kinyua')
   end
   context 'create instance of class Author' do
     it 'takes three parameters and return a Label object' do
@@ -18,7 +18,7 @@ describe Author do
   end
   context 'When adding a game into an author' do
     it 'adds the game into a specific author' do
-      @author.add_author(@game)
+      @author.add_item(@game)
       expect(@author.items).to include(@game)
       expect(@author.items.length).to eq(1)
     end
